split JsonExtractTests by category to fit iters=100 heap
The repeat-changed-tests CI runner OOM'd JsonExtractTests at
-Dtests.iters=100 with the default 512m test heap. Cause: Gradle's
JUnit descriptor tree at testRunStarted — ~80 parameterized suppliers
× anyNullIsNull doubling × 30 inherited test methods × 100 iters =
~480k Description objects, exceeding the heap before any test ran.
No leak in our code; descriptor-tree scale.

Splitting JsonExtractTests by category so each test class fits its
fork's heap budget:

- JsonExtractTests — core value extraction (type coverage, scalar
  values, nested objects, arrays, constant path, null result).
- JsonExtractPathSyntaxTests (new) — bracket notation, $ prefix
  variants, escaped chars, Unicode keys.
- JsonExtractWarningTests (new) — warning-emitting cases (invalid
  JSON, missing path, array bounds, unsupported JSONPath syntax,
  null-source).
- JsonExtractStaticTests (new, extends ESTestCase, follows the
  established *StaticTests pattern) — standalone non-parameterized:
  large/deep input, XContent encoding coverage, randomized fuzz,
  evaluator routing, null-source warning function-level contract.

Each parameterized class still covers all declared @Param.type
values via a SOURCE null-source supplier whose warning expectations
also skip the framework's null-injection tests via assumeTrue.

Refs #149521
